Change extract frametimes to accomodate MEA60
Previously, the binary file name for the analog channel containing
the pulses was hard coded. However, for recordings from 60-electrode
array requires using a different file name as well as changing
other parameters. Now all of this is handled automatically.

Also the conversion factor "microvoltsperADunit" was inaccurate for
252 as well, now it is closer to the real values. This does not
affect the threshold crossings in a significant way but the change
introduced a difference of 0.3% in the timing values compared to
previous version. I do not expect any changes to the actual analyses.