Outsource experiment paths into config file
The configuration is formatted as JSON file to allow any future/nested
settings to be added. The file is excluded via .gitignore. If the file
does not exist, it will be created with empty entries. Parsing
errors will raise an exception with specified line and column
number of the syntax error.