Fix issue with untracked configuration file
Since config.json cannot be present in the repository with local changes
reliably ignored across branches, it is now removed from the repository
and ignored locally. Again, it will be created on first use, but this
time with invalid value ('None') for the 'root_experiment_dir' key. When
kept, this will raise an exception to prevent undefined behavior and
will force adjusting the configuration file until valid.