Implement default and user configuration
The default configuration file contains annotative comments (not typical
for the JSON format) to guide when copying individual settings into a
local, custom configuration file in the home directory.
When accessing settings, the user config is parsed over the default
config. After loading the config from disk it is cached for performance.

Fix #4