Use frametime averages for on-off steps trial periods
Previously, the settings from the parameters file was used for
determining the length of each period in a trial (preframe, on, off).
This was quite inaccurate (up to 10ms).

Each trial was aligned to the end of previous one, so the cumulative
error was reset at each trial. However, the problem was clear for the
off period, which is the last one; some primate cells seemed to respond
before the screen actually changed.

This does not seem to have affected the on-off indices of mouse cells
too much, as they are slower.