Add smaller version for texture generated with loops
The loop version generates 800x800 but the rest of the plotting and
analysis functions for the texture in OMB object require a smaller
version (i.e 200x200 if bgstixel=4).

To be able to compare the output of the numpy version of the texture
and the loop version, now loop version also creates a smaller image.