Use pyglmnet for finding the GQM filters
The optimization problem of fitting the quadratic model to the cells can
be reformulated so that the quadratic term (sT Q s) is not calculated
each time Q is changed. Using (B sT Q s), the quadratic term can be
calculated once and B can be optimized. When flattened B is fed into the
algorithm as a linear term.

The package pyglmnet allows using regularization and cross-validation.